+++ odoo/local-src/mail_company_domain/__manifest__.py
@@ -1,0 +1,16 @@
+# -*- coding: utf-8 -*-
+# Copyright 2017 Camptocamp SA
+# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html)
+
+{'name': 'E-mail Domain per Company',
+ 'version': '10.0.1.0.0',
+ 'author': 'Camptocamp,Odoo Community Association (OCA)',
+ 'license': 'AGPL-3',
+ 'category': 'category',
+ 'depends': ['mail',
+             ],
+ 'website': 'https://www.camptocamp.com',
+ 'data': ['security/ir.model.access.csv',
+          ],
+ 'installable': True,
+ }