+++ odoo/local-src/mail_company_domain/models/__init__.py
@@ -1,0 +1,9 @@
+# -*- coding: utf-8 -*-
+
+from . import ir_config_parameter
+from . import ir_mail_server
+from . import mail_catchall_domain
+from . import mail_alias
+from . import mail_message
+from . import mail_mail
+from . import mail_thread